requirements.txt
asgiref==3.8.1
certifi==2024.2.2
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
class-registry==2.1.2
cryptography==42.0.5
defusedxml==0.8.0rc2
Django==4.2.11
django-filter==24.2
django-templated-mail==1.1.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
djoser==2.2.2
filters==1.3.2
flake8==6.0.0
flake8-isort==6.0.0
gunicorn==20.1.0
idna==3.6
isort==5.13.2
mccabe==0.7.0
oauthlib==3.2.2
pillow==10.3.0
psycopg2==2.9.9
pycodestyle==2.10.0
pycparser==2.22
pyflakes==3.0.1
PyJWT==2.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python3-openid==3.2.0
pytz==2024.1
regex==2023.12.25
reportlab==4.1.0
requests==2.31.0
requests-oauthlib==2.0.0
six==1.16.0
social-auth-app-django==5.4.0
social-auth-core==4.5.3
sqlparse==0.4.4
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
webcolors==1.13
